+++ src/compas_timber/connections/__init__.py
@@ -25,9 +25,6 @@ from .oligina import OliGinaJoint
 from .utilities import beam_ref_side_incidence
 from .utilities import beam_ref_side_incidence_with_vector
 from .utilities import point_centerline_towards_joint
-from .wall_joint import WallJoint
-from .wall_joint import InterfaceLocation
-from .wall_joint import InterfaceRole
 from .plate_joint import PlateJoint
 from .slab_joint import SlabJoint
 from .plate_butt_joint import PlateButtJoint
@@ -75,9 +72,6 @@ __all__ = [
     "beam_ref_side_incidence",
     "beam_ref_side_incidence_with_vector",
     "point_centerline_towards_joint",
-    "WallJoint",
-    "InterfaceLocation",
-    "InterfaceRole",
     "PlateJoint",
     "SlabJoint",
     "PlateButtJoint",