+++ src/compas_timber/elements/slab.py
@@ -1,6 +1,7 @@
 from compas.geometry import Point
 from compas.geometry import Polyline
 from compas.geometry import Transformation
+from compas.geometry import Frame
 from compas_model.elements import Element
 from compas_model.elements import reset_computed
 
@@ -31,7 +32,7 @@ class SlabType(object):
 class Slab(PlateGeometry, Element):
     """Represents a timber slab element (wall, floor, roof, etc.).
 
-    Serves as container for beams, joints and other related elements and groups them together to form a slab.
+    Serves as container for beams, plates, and other related elements and groups them together to form a slab.
     A slab is often a single unit of prefabricated timber element.
     It is often referred to as an enveloping body.
 
@@ -45,9 +46,9 @@ class Slab(PlateGeometry, Element):
         Width of the slab.
     thickness : float
         Thickness of the slab.
-    outline_a : :class:`~compas.geometry.Polyline`, optional
+    local_outline_a: :class:`~compas.geometry.Polyline`, optional
         A polyline representing the principal outline of this slab.
-    outline_b : :class:`~compas.geometry.Polyline`, optional
+    local_outline_b: :class:`~compas.geometry.Polyline`, optional
         A polyline representing the associated outline of this slab.
     openings : list[:class:`~compas_timber.elements.Opening`], optional
         A list of Opening objects representing openings in this slab.
@@ -74,14 +75,6 @@ class Slab(PlateGeometry, Element):
         List of interfaces associated with this slab.
     attributes : dict
         Dictionary of additional attributes.
-    is_slab : bool
-        Always True for slabs.
-    is_wall : bool
-        False for base Slab class.
-    is_floor : bool
-        False for base Slab class.
-    is_roof : bool
-        False for base Slab class.
     is_group_element : bool
         Always True for slabs as they can contain other elements.
 
@@ -97,24 +90,24 @@ class Slab(PlateGeometry, Element):
         data["name"] = self.name
         return data
 
-    def __init__(self, transformation, length, width, thickness, local_outline_a=None, local_outline_b=None, openings=None, name=None, **kwargs):
-        Element.__init__(self, transformation=transformation, **kwargs)
+    def __init__(self, frame, length, width, thickness, local_outline_a=None, local_outline_b=None, openings=None, type=None,name=None, **kwargs):
+        Element.__init__(self, transformation=Transformation.from_frame(frame) if frame else Transformation(), name=name, **kwargs)
         local_outline_a = local_outline_a or Polyline([Point(0, 0, 0), Point(length, 0, 0), Point(length, width, 0), Point(0, width, 0), Point(0, 0, 0)])
         local_outline_b = local_outline_b or Polyline([Point(p[0], p[1], thickness) for p in local_outline_a.points])
         PlateGeometry.__init__(self, local_outline_a, local_outline_b, openings=openings)
         self.length = length
         self.width = width
         self.height = thickness
-        self.name = name or "Slab"
         self.interfaces = []
+        self.type = type or SlabType.GENERIC
         self.attributes = {}
         self.attributes.update(kwargs)
 
     def __repr__(self):
-        return "Slab(name={}, {}, {}, {:.3f})".format(self.name, self.transformation, self.outline_a, self.thickness)
+        return "Slab(name={}, {}, {}, {:.3f})".format(self.name, Frame.from_transformation(self.transformation), self.outline_a, self.thickness)
 
     def __str__(self):
-        return "Slab(name={}, {}, {}, {:.3f})".format(self.name, self.transformation, self.outline_a, self.thickness)
+        return "Slab(name={}, {}, {}, {:.3f})".format(self.name, Frame.from_transformation(self.transformation), self.outline_a, self.thickness)
 
     @reset_computed
     def reset(self):
@@ -142,58 +135,7 @@ class Slab(PlateGeometry, Element):
             self.interfaces = [i for i in self.interfaces if i not in interfaces]
 
     @property
-    def is_slab(self):
-        """Check if this element is a slab.
-
-        Returns
-        -------
-        bool
-            Always True for slabs.
-        """
-        return True
-
-    @property
-    def is_wall(self):
-        """Check if this element is a wall.
-
-        Returns
-        -------
-        bool
-            False for the base Slab class.
-        """
-        return False
-
-    @property
-    def is_floor(self):
-        """Check if this element is a floor.
-
-        Returns
-        -------
-        bool
-            False for the base Slab class.
-        """
-        return False
-
-    @property
-    def is_roof(self):
-        """Check if this element is a roof.
-
-        Returns
-        -------
-        bool
-            False for the base Slab class.
-        """
-        return False
-
-    @property
     def is_group_element(self):
-        """Check if this element can be used as a container for other elements.
-
-        Returns
-        -------
-        bool
-            Always True for slabs as they can contain other elements.
-        """
         return True
 
     def compute_modeltransformation(self):
@@ -237,5 +179,4 @@ class Slab(PlateGeometry, Element):
         args = PlateGeometry.get_args_from_outlines(outline_a, outline_b, openings)
         PlateGeometry._check_outlines(args["local_outline_a"], args["local_outline_b"])
         kwargs.update(args)
-        kwargs["transformation"] = Transformation.from_frame(args["frame"])
         return cls(**kwargs)